+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/model/exporter/AbstractModelExporter.java
@@ -19,6 +19,7 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.NotDirectoryException;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.text.MessageFormat;
@@ -36,9 +37,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import ru.taximaxim.codekeeper.core.Consts;
+import ru.taximaxim.codekeeper.core.DatabaseType;
 import ru.taximaxim.codekeeper.core.PgCodekeeperException;
 import ru.taximaxim.codekeeper.core.PgDiffUtils;
 import ru.taximaxim.codekeeper.core.UnixPrintWriter;
+import ru.taximaxim.codekeeper.core.WorkDirs;
 import ru.taximaxim.codekeeper.core.fileutils.FileUtils;
 import ru.taximaxim.codekeeper.core.localizations.Messages;
 import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
@@ -127,7 +130,24 @@ public abstract class AbstractModelExporter {
         writeProjVersion(outDir.resolve(Consts.FILENAME_WORKING_DIR_MARKER));
     }
 
-    protected abstract void createOutDir() throws IOException;
+    protected void createOutDir() throws IOException {
+        if (Files.exists(outDir)) {
+            if (!Files.isDirectory(outDir)) {
+                throw new NotDirectoryException(outDir.toString());
+            }
+
+            for (String subdirName : WorkDirs.getDirectoryNames(getDatabaseType())) {
+                if (Files.exists(outDir.resolve(subdirName))) {
+                    throw new DirectoryException(
+                            MessageFormat.format("Output directory already contains {0} directory.", subdirName));
+                }
+            }
+        } else {
+            Files.createDirectories(outDir);
+        }
+    }
+
+    protected abstract DatabaseType getDatabaseType();
 
     public void exportPartial() throws IOException, PgCodekeeperException {
         if (oldDb == null) {
@@ -268,6 +288,38 @@ public abstract class AbstractModelExporter {
             throw new IllegalArgumentException(Messages.DatabaseType_unsupported_type + st.getDbType());
         }
         return exporter.getRelativeFilePath(st, true);
+    }
+
+    public static void exportPartial(DatabaseType dbType, Path dirExport, PgDatabase dbNew, PgDatabase dbOld,
+            Collection<TreeElement> changedObjects, String encoding) throws IOException, PgCodekeeperException {
+        AbstractModelExporter exporter;
+        switch (dbType) {
+        case MS:
+            exporter = new MsModelExporter(dirExport, dbNew, dbOld, changedObjects, encoding);
+            break;
+        case PG:
+            exporter = new ModelExporter(dirExport, dbNew, dbOld, changedObjects, encoding);
+            break;
+        default:
+            throw new IllegalArgumentException(Messages.DatabaseType_unsupported_type + dbType);
+        }
+        exporter.exportPartial();
+    }
+
+    public static void exportFull(DatabaseType dbType, Path dirExport, PgDatabase dbNew, String encoding)
+            throws IOException {
+        AbstractModelExporter exporter;
+        switch (dbType) {
+        case PG:
+            exporter = new ModelExporter(dirExport, dbNew, encoding);
+            break;
+        case MS:
+            exporter = new MsModelExporter(dirExport, dbNew, encoding);
+            break;
+        default:
+            throw new IllegalArgumentException(Messages.DatabaseType_unsupported_type + dbType);
+        }
+        exporter.exportFull();
     }
 }