+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/model/exporter/ModelExporter.java
@@ -15,15 +15,12 @@
  *******************************************************************************/
 package ru.taximaxim.codekeeper.core.model.exporter;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.NotDirectoryException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.text.MessageFormat;
 import java.util.Collection;
 
-import ru.taximaxim.codekeeper.core.Consts;
+import ru.taximaxim.codekeeper.core.DatabaseType;
+import ru.taximaxim.codekeeper.core.WorkDirs;
 import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
 import ru.taximaxim.codekeeper.core.model.difftree.TreeElement;
 import ru.taximaxim.codekeeper.core.schema.PgDatabase;
@@ -41,33 +38,13 @@ public class ModelExporter extends AbstractModelExporter {
     }
 
     @Override
-    protected void createOutDir() throws IOException {
-        if (Files.exists(outDir)) {
-            if (!Files.isDirectory(outDir)) {
-                throw new NotDirectoryException(outDir.toString());
-            }
-
-            for (Consts.WORK_DIR_NAMES subdirName : Consts.WORK_DIR_NAMES.values()) {
-                if (Files.exists(outDir.resolve(subdirName.name()))) {
-                    throw new DirectoryException(MessageFormat.format(
-                            "Output directory already contains {0} directory.",
-                            subdirName));
-                }
-            }
-        } else {
-            Files.createDirectories(outDir);
-        }
-    }
-
-    @Override
     protected Path getRelativeFilePath(PgStatement st, boolean addExtension) {
         return getRelativeFilePath(st, Paths.get(""), addExtension);
     }
 
     static Path getRelativeFilePath(PgStatement st, Path baseDir, boolean addExtension) {
         PgStatement parentSt = st.getParent();
-        String parentExportedFileName = parentSt == null ?
-                null : getExportedFilename(parentSt);
+        String parentExportedFileName = parentSt == null ? null : getExportedFilename(parentSt);
 
         Path path = baseDir.resolve("SCHEMA");
         DbObjType type = st.getStatementType();
@@ -78,11 +55,8 @@ public class ModelExporter extends AbstractModelExporter {
         case USER_MAPPING:
         case CAST:
         case EVENT_TRIGGER:
-            path = baseDir.resolve(type.name());
-            break;
-
         case FOREIGN_DATA_WRAPPER:
-            path = baseDir.resolve("FDW");
+            path = baseDir.resolve(WorkDirs.getDirectoryNameForType(st.getDbType(), type));
             break;
 
         case SCHEMA:
@@ -117,14 +91,18 @@ public class ModelExporter extends AbstractModelExporter {
         case POLICY:
         case COLUMN:
             st = parentSt;
-            schemaName = AbstractModelExporter.getExportedFilename(parentSt.getParent());
+            schemaName = getExportedFilename(parentSt.getParent());
             path = path.resolve(schemaName).resolve(parentSt.getStatementType().name());
             break;
         default:
             break;
         }
 
-        return path.resolve(addExtension ?
-                getExportedFilenameSql(st) : getExportedFilename(st));
+        return path.resolve(addExtension ? getExportedFilenameSql(st) : getExportedFilename(st));
+    }
+
+    @Override
+    protected DatabaseType getDatabaseType() {
+        return DatabaseType.PG;
     }
 }