+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/pgdbproject/NewProjWizard.java
@@ -33,7 +33,6 @@ import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.IExecutableExtension;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.jface.operation.IRunnableWithProgress;
-import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.jface.viewers.ArrayContentProvider;
 import org.eclipse.jface.viewers.ComboViewer;
 import org.eclipse.jface.viewers.IStructuredSelection;
@@ -67,12 +66,11 @@ import ru.taximaxim.codekeeper.core.loader.JdbcRunner;
 import ru.taximaxim.codekeeper.core.schema.PgDatabase;
 import ru.taximaxim.codekeeper.ui.Activator;
 import ru.taximaxim.codekeeper.ui.Log;
+import ru.taximaxim.codekeeper.ui.ProjectIcon;
 import ru.taximaxim.codekeeper.ui.UIConsts;
 import ru.taximaxim.codekeeper.ui.UIConsts.DB_BIND_PREF;
-import ru.taximaxim.codekeeper.ui.UIConsts.FILE;
 import ru.taximaxim.codekeeper.ui.UIConsts.PREF_PAGE;
 import ru.taximaxim.codekeeper.ui.UIConsts.PROJ_PREF;
-import ru.taximaxim.codekeeper.ui.UIConsts.WIZARD;
 import ru.taximaxim.codekeeper.ui.UIConsts.WORKING_SET;
 import ru.taximaxim.codekeeper.ui.dbstore.DbInfo;
 import ru.taximaxim.codekeeper.ui.dbstore.DbMenuStorePicker;
@@ -88,15 +86,13 @@ implements IExecutableExtension, INewWizard {
     private PageRepo pageRepo;
     private PageDb pageDb;
 
-    private DatabaseType dbType;
     private IConfigurationElement config;
     private IWorkbench workbench;
     private IStructuredSelection selection;
 
     public NewProjWizard() {
         setWindowTitle(Messages.newProjWizard_new_pg_db_project);
-        setDefaultPageImageDescriptor(ImageDescriptor.createFromURL(
-                Activator.getDefault().getBundle().getResource(FILE.ICONAPPWIZ)));
+        setDefaultPageImageDescriptor(Activator.getRegisteredDescriptor(ProjectIcon.APP_WIZ));
         setNeedsProgressMonitor(true);
     }
 
@@ -105,12 +101,13 @@ implements IExecutableExtension, INewWizard {
         // page names shouldn't be localized, use page titles instead
         pageRepo = new PageRepo("main", selection); //$NON-NLS-1$
         addPage(pageRepo);
-        pageDb = new PageDb(dbType, "schema"); //$NON-NLS-1$
+        pageDb = new PageDb("schema"); //$NON-NLS-1$
         addPage(pageDb);
     }
 
     @Override
     public boolean performFinish() {
+        DatabaseType dbType = pageDb.getDbType();
         PgDbProject props = null;
         boolean initSuccess = false;
         try {
@@ -198,7 +195,7 @@ implements IExecutableExtension, INewWizard {
         } else if (dbinfo != null) {
             src = DbSource.fromDbInfo(dbinfo, forceUnixNewlines, charset, timezone, props.getProject());
         } else if ((dump = pageDb.getDumpPath()) != null) {
-            src = DbSource.fromFile(forceUnixNewlines, dump, charset, dbType, props.getProject());
+            src = DbSource.fromFile(forceUnixNewlines, dump, charset, pageDb.getDbType(), props.getProject());
         } else {
             // should be prevented by page completion state
             throw new IllegalStateException(Messages.initProjectFromSource_init_request_but_no_schema_source);
@@ -211,10 +208,6 @@ implements IExecutableExtension, INewWizard {
     public void setInitializationData(IConfigurationElement config,
             String propertyName, Object data) throws CoreException {
         this.config = config;
-        this.dbType = WIZARD.NEW_PROJECT_WIZARD.equals(config.getAttribute("id")) ? DatabaseType.PG : DatabaseType.MS; //$NON-NLS-1$
-        if (dbType == DatabaseType.MS) {
-            setWindowTitle(Messages.NewProjWizard_new_ms_project);
-        }
     }
 
     @Override
@@ -265,7 +258,7 @@ class PageDb extends WizardPage {
     private static final String QUERY_TZ = "SELECT name, setting FROM pg_catalog.pg_file_settings" //$NON-NLS-1$
             + " WHERE pg_catalog.lower(name) = 'timezone' AND applied AND error IS NULL"; //$NON-NLS-1$
 
-    private final DatabaseType dbType;
+    private DatabaseType dbType;
 
     private Button btnInit;
     private Button btnBind;
@@ -274,6 +267,7 @@ class PageDb extends WizardPage {
     private ComboViewer timezoneCombo;
     private ComboViewer charsetCombo;
     private CLabel lblWarnPosix;
+    private Label timeZoneLabel;
 
 
     public DbInfo getDbInfo() {
@@ -300,9 +294,12 @@ class PageDb extends WizardPage {
         return timezoneCombo.getCombo().getText();
     }
 
-    PageDb(DatabaseType dbType, String pageName) {
+    public DatabaseType getDbType( ) {
+        return dbType;
+    }
+
+    PageDb(String pageName) {
         super(pageName, pageName, null);
-        this.dbType = dbType;
 
         setTitle(Messages.NewProjWizard_proj_init);
         setDescription(Messages.NewProjWizard_proj_init_src);
@@ -312,6 +309,10 @@ class PageDb extends WizardPage {
     public void createControl(final Composite parent) {
         Composite container = new Composite(parent, SWT.NONE);
         container.setLayout(new GridLayout(3, false));
+
+        new Label(container, SWT.NONE).setText(Messages.NewProjWizard_select_project_type);
+        var cmbDbType = new ComboViewer(container);
+        cmbDbType.getCombo().setLayoutData(new GridData(SWT.FILL, SWT.NONE, true, false, 2, 1));
 
         //init db block
         Group group = new Group(container, SWT.NONE);
@@ -343,11 +344,10 @@ class PageDb extends WizardPage {
         new Label(source, SWT.NONE).setText(Messages.DbStorePicker_db_connection);
 
         storePicker = new DbMenuStorePicker(source, true, false);
-        storePicker.filter(dbType);
         storePicker.addSelectionListener(this::modifyButtons);
 
         Button btnEditStore = new Button(source, SWT.PUSH);
-        btnEditStore.setImage(Activator.getRegisteredImage(FILE.ICONEDIT));
+        btnEditStore.setImage(Activator.getRegisteredImage(ProjectIcon.EDIT));
         btnEditStore.addSelectionListener(new SelectionAdapter() {
 
             @Override
@@ -367,49 +367,64 @@ class PageDb extends WizardPage {
         charsetCombo.setSelection(new StructuredSelection(Consts.UTF_8));
 
         //time zones
-        if (dbType == DatabaseType.PG) {
-            new Label(container, SWT.NONE).setText(Messages.NewProjWizard_select_time_zone);
+        timeZoneLabel = new Label(container, SWT.NONE);
+        timeZoneLabel.setText(Messages.NewProjWizard_select_time_zone);
+        timezoneCombo = new ComboViewer(container, SWT.DROP_DOWN);
+        timezoneCombo.getCombo().setLayoutData(new GridData(SWT.FILL, SWT.DEFAULT, true, false));
+        timezoneCombo.setContentProvider(ArrayContentProvider.getInstance());
+        timezoneCombo.setInput(UIConsts.TIME_ZONES);
+        timezoneCombo.setSelection(new StructuredSelection(Consts.UTC));
+        timezoneCombo.getCombo().addModifyListener(e -> timeZoneWarn());
 
-            timezoneCombo = new ComboViewer(container, SWT.DROP_DOWN);
-            timezoneCombo.getCombo().setLayoutData(new GridData(SWT.FILL, SWT.DEFAULT, true, false));
-            timezoneCombo.setContentProvider(ArrayContentProvider.getInstance());
-            timezoneCombo.setInput(UIConsts.TIME_ZONES);
-            timezoneCombo.setSelection(new StructuredSelection(Consts.UTC));
-            timezoneCombo.getCombo().addModifyListener(e -> timeZoneWarn());
+        btnGetTz = new Button(container, SWT.PUSH);
+        btnGetTz.setText(Messages.NewProjWizard_get_from_db);
+        btnGetTz.setEnabled(storePicker.getDbInfo() != null);
+        btnGetTz.addSelectionListener(new SelectionAdapter() {
 
-            btnGetTz = new Button(container, SWT.PUSH);
-            btnGetTz.setText(Messages.NewProjWizard_get_from_db);
-            btnGetTz.setEnabled(storePicker.getDbInfo() != null);
-            btnGetTz.addSelectionListener(new SelectionAdapter() {
-
-                @Override
-                public void widgetSelected(SelectionEvent e) {
-                    DbInfo dbinfo = getDbInfo();
-                    if (dbinfo == null) {
-                        return;
-                    }
-                    try {
-                        TimeZoneProgress progress = new TimeZoneProgress(dbinfo);
-                        getContainer().run(true, true, progress);
-                        String timezone = progress.timezone;
-                        timezoneCombo.getCombo().setText(timezone == null ? Consts.UTC : timezone);
-                    } catch (InterruptedException ex) {
-                        // cancelled
-                    } catch (InvocationTargetException ex) {
-                        ExceptionNotifier.notifyDefault(Messages.NewProjWizard_error_tz_query, ex);
-                    }
+            @Override
+            public void widgetSelected(SelectionEvent e) {
+                DbInfo dbinfo = getDbInfo();
+                if (dbinfo == null) {
+                    return;
                 }
-            });
+                try {
+                    TimeZoneProgress progress = new TimeZoneProgress(dbinfo);
+                    getContainer().run(true, true, progress);
+                    String timezone = progress.timezone;
+                    timezoneCombo.getCombo().setText(timezone == null ? Consts.UTC : timezone);
+                } catch (InterruptedException ex) {
+                    // cancelled
+                } catch (InvocationTargetException ex) {
+                    ExceptionNotifier.notifyDefault(Messages.NewProjWizard_error_tz_query, ex);
+                }
+            }
+        });
 
-            lblWarnPosix = new CLabel(container, SWT.NONE);
-            lblWarnPosix.setImage(Activator.getEclipseImage(ISharedImages.IMG_OBJS_WARN_TSK));
-            lblWarnPosix.setText(Messages.ProjectProperties_posix_is_used_warn);
-            GridData gd = new GridData(SWT.FILL, SWT.DEFAULT, false, false, 3, 1);
-            gd.exclude = true;
-            lblWarnPosix.setLayoutData(gd);
-        }
+        cmbDbType.setContentProvider(ArrayContentProvider.getInstance());
+        cmbDbType.setLabelProvider(UIConsts.DATABASE_TYPE_PROVIDER);
+        cmbDbType.setInput(DatabaseType.values());
+        cmbDbType.addSelectionChangedListener(e -> {
+            StructuredSelection sel = (StructuredSelection) e.getSelection();
+            dbType = (DatabaseType) sel.getFirstElement();
+            storePicker.filter(dbType);
+            updateVisibleTimeZone();
+        });
+        cmbDbType.setSelection(new StructuredSelection(DatabaseType.PG));
+
+        lblWarnPosix = new CLabel(container, SWT.NONE);
+        lblWarnPosix.setImage(Activator.getEclipseImage(ISharedImages.IMG_OBJS_WARN_TSK));
+        lblWarnPosix.setText(Messages.ProjectProperties_posix_is_used_warn);
+        GridData gd = new GridData(SWT.FILL, SWT.DEFAULT, false, false, 3, 1);
+        gd.exclude = true;
+        lblWarnPosix.setLayoutData(gd);
 
         setControl(container);
+    }
+
+    private void updateVisibleTimeZone() {
+        timezoneCombo.getCombo().setVisible(dbType == DatabaseType.PG);
+        timeZoneLabel.setVisible(dbType == DatabaseType.PG);
+        btnGetTz.setVisible(dbType == DatabaseType.PG);
     }
 
     private void modifyButtons() {
@@ -465,9 +480,9 @@ class PageDb extends WizardPage {
         public void run(IProgressMonitor monitor)
                 throws InvocationTargetException, InterruptedException {
 
-            JdbcConnector connector = new JdbcConnector(dbinfo.getDbHost(), dbinfo.getDbPort(),
-                    dbinfo.getDbUser(), dbinfo.getDbPass(), dbinfo.getDbName(), dbinfo.getProperties(),
-                    dbinfo.isReadOnly(), Consts.UTC);
+            JdbcConnector connector = JdbcConnector.getJdbcConnector(DatabaseType.PG,
+                    dbinfo.getDbHost(), dbinfo.getDbPort(), dbinfo.getDbUser(), dbinfo.getDbPass(), dbinfo.getDbName(),
+                    dbinfo.getProperties(), dbinfo.isReadOnly(), Consts.UTC, false, null);
 
             try (Connection connection = connector.getConnection();
                     Statement st = connection.createStatement();